Fake News Detection using BERT/requirements.txt
transformers[torch] 
accelerate -U
torch 
transformers
pandas
logging
sklearn.model_selection